Resolve vehicle icons from one map instead of six copies
